+++ web/src/components/Spots/PostDetail/PostDetail.module.scss
